fix: correct dbtypes import paths and refine domain/transport layer
Fix broken import paths referencing internal/dbtypes instead of
internal/db/dbtypes across domain and services. Includes refinements
to repositories, handlers, scraper, feed UI, and updated pipeline docs.

+++ backend/internal/middleware/auth.go
@@ -69,12 +69,12 @@ func OptionalAuthMiddleware(authService *services.AuthService) gin.HandlerFunc {
 	}
 }
 
-func GetUserID(c *gin.Context) (int, bool) {
+func GetUserID(c *gin.Context) (int64, bool) {
 	userID, exists := c.Get("user_id")
 	if !exists {
 		return 0, false
 	}
-	id, ok := userID.(int)
+	id, ok := userID.(int64)
 	return id, ok
 }
